Refactor: #249 AgreementBottomSheet 컴포넌트 분리
